server/requirements.txt
asgiref==3.4.1
astroid==2.12.13
autopep8==2.0.1
colorama==0.4.6
dill==0.3.6
Django==3.2.16
django-cors-headers==3.10.1
djangorestframework==3.13.0
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
flake8-django==1.1.5
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==2.6.2
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pylint==2.15.9
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.1.1
wrapt==1.14.1
